+++ src/main/java/fll/web/scoreEntry/PerformanceRunsEndpoint.java
@@ -13,6 +13,8 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -22,9 +24,11 @@ import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
 
 import javax.sql.DataSource;
 
+import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.text.StringEscapeUtils;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -36,7 +40,13 @@ import static org.checkerframework.checker.nullness.util.NullnessUtil.castNonNul
 
 import fll.Team;
 import fll.Tournament;
+import fll.TournamentTeam;
 import fll.Utilities;
+import fll.db.Queries;
+import fll.db.TournamentParameters;
+import fll.scheduler.PerformanceTime;
+import fll.scheduler.TeamScheduleInfo;
+import fll.scheduler.TournamentSchedule;
 import fll.util.FLLInternalException;
 import fll.util.FLLRuntimeException;
 import fll.web.ApplicationAttributes;
@@ -78,7 +88,8 @@ public class PerformanceRunsEndpoint {
       final DataSource datasource = ApplicationAttributes.getDataSource(application);
       try (Connection connection = datasource.getConnection()) {
         final Collection<UnverifiedRunData> unverifiedData = getUnverifiedRunData(connection);
-        final PerformanceRunData runData = new PerformanceRunData(unverifiedData);
+        final List<SelectTeamData> teamSelectData = getTeamSelectData(connection);
+        final PerformanceRunData runData = new PerformanceRunData(teamSelectData, unverifiedData);
         final ObjectMapper jsonMapper = Utilities.createJsonMapper();
         final String msg = jsonMapper.writeValueAsString(runData);
         sendToClient(session, msg);
@@ -140,13 +151,15 @@ public class PerformanceRunsEndpoint {
    */
   public static void notifyToUpdate(Connection connection) throws SQLException {
     final Collection<UnverifiedRunData> unverifiedData = getUnverifiedRunData(connection);
-    sendData(unverifiedData);
+    final List<SelectTeamData> teamSelectData = getTeamSelectData(connection);
+    sendData(teamSelectData, unverifiedData);
   }
 
-  private static void sendData(final Collection<UnverifiedRunData> unverifiedData) {
+  private static void sendData(final List<SelectTeamData> teamSelectData,
+                               final Collection<UnverifiedRunData> unverifiedData) {
     THREAD_POOL.execute(() -> {
       try {
-        final PerformanceRunData runData = new PerformanceRunData(unverifiedData);
+        final PerformanceRunData runData = new PerformanceRunData(teamSelectData, unverifiedData);
         final ObjectMapper jsonMapper = Utilities.createJsonMapper();
         final String msg = jsonMapper.writeValueAsString(runData);
 
@@ -183,6 +196,152 @@ public class PerformanceRunsEndpoint {
       client.getBasicRemote().sendText(message);
     } catch (final IllegalStateException e) {
       throw new IOException("Illegal state exception writing to client", e);
+    }
+  }
+
+  private static List<SelectTeamData> getTeamSelectData(final Connection connection) throws SQLException {
+    final Tournament tournament = Tournament.getCurrentTournament(connection);
+
+    // get latest performance round for each team
+    final Map<Integer, Integer> maxRunNumbers = Collections.unmodifiableMap(getMaxRunNumbers(connection, tournament));
+
+    final @Nullable TournamentSchedule schedule;
+    if (TournamentSchedule.scheduleExistsInDatabase(connection, tournament.getTournamentID())) {
+      schedule = new TournamentSchedule(connection, tournament.getTournamentID());
+    } else {
+      schedule = null;
+    }
+
+    final List<String> unfinishedBrackets = Collections.unmodifiableList(Playoff.getUnfinishedPlayoffBrackets(connection,
+                                                                                                              tournament.getTournamentID()));
+
+    final List<PerformanceRunsEndpoint.SelectTeamData> teamSelectData = Queries.getTournamentTeams(connection,
+                                                                                                   tournament.getTournamentID())
+                                                                               .values().stream() //
+                                                                               .filter(team -> !team.isInternal()) //
+                                                                               .map(team -> teamToTeamSelectData(connection,
+                                                                                                                 tournament,
+                                                                                                                 maxRunNumbers,
+                                                                                                                 schedule,
+                                                                                                                 unfinishedBrackets,
+                                                                                                                 team)) //
+                                                                               .collect(Collectors.toList());
+
+    return teamSelectData;
+  }
+
+  /**
+   * Find the maximum run number all teams. Does not ignore unverified
+   * scores. If a team is not in the result, then the next run number is 1.
+   * 
+   * @param connection database connection
+   * @param tournament the tournament to check
+   * @return the maximum run number recorded for each team
+   * @throws SQLException on a database error
+   */
+  private static Map<Integer, Integer> getMaxRunNumbers(final Connection connection,
+                                                        final Tournament tournament)
+      throws SQLException {
+
+    final Map<Integer, Integer> maxRunNumbers = new HashMap<>();
+    try (
+        PreparedStatement prep = connection.prepareStatement("SELECT TeamNumber, COUNT(TeamNumber) as max_run FROM Performance"
+            + " WHERE Tournament = ?"
+            + " GROUP BY TeamNumber")) {
+      prep.setInt(1, tournament.getTournamentID());
+      try (ResultSet rs = prep.executeQuery()) {
+        while (rs.next()) {
+          final int teamNumber = rs.getInt("TeamNumber");
+          final int runNumber = rs.getInt("max_run");
+          maxRunNumbers.put(teamNumber, runNumber);
+        }
+      } // result set
+    } // prepared statement
+
+    return maxRunNumbers;
+  }
+
+  private static PerformanceRunsEndpoint.SelectTeamData teamToTeamSelectData(Connection connection,
+                                                                             final Tournament tournament,
+                                                                             final Map<Integer, Integer> maxRunNumbers,
+                                                                             final @Nullable TournamentSchedule schedule,
+                                                                             final List<String> unfinishedBrackets,
+                                                                             final TournamentTeam team) {
+    final int nextRunNumber = maxRunNumbers.getOrDefault(team.getTeamNumber(), 0)
+        + 1;
+    final @Nullable PerformanceTime nextPerformance = getNextPerformance(schedule, team.getTeamNumber(), nextRunNumber);
+    if (null != nextPerformance) {
+      return new PerformanceRunsEndpoint.SelectTeamData(team, nextPerformance, nextRunNumber);
+    } else {
+
+      final String nextTableAndSide = getNextTableAndSide(connection, tournament, team.getTeamNumber(), nextRunNumber);
+
+      try {
+        final String playoffBracketName = Playoff.getPlayoffBracketsForTeam(connection, team.getTeamNumber()).stream() //
+                                                 .filter(b -> unfinishedBrackets.contains(b)) //
+                                                 .findFirst() //
+                                                 .orElse("");
+        final String runNumberDisplay;
+        if (!playoffBracketName.isEmpty()) {
+          final int playoffRound = Playoff.getPlayoffRound(connection, tournament.getTournamentID(), playoffBracketName,
+                                                           nextRunNumber);
+
+          final int playoffMatch = Playoff.getPlayoffMatch(connection, tournament.getTournamentID(), playoffBracketName,
+                                                           nextRunNumber);
+          runNumberDisplay = String.format("%s P%d M%d", playoffBracketName, playoffRound, playoffMatch);
+        } else {
+          runNumberDisplay = String.valueOf(nextRunNumber);
+        }
+        return new PerformanceRunsEndpoint.SelectTeamData(team, nextTableAndSide, nextRunNumber, runNumberDisplay);
+      } catch (final SQLException e) {
+        LOGGER.warn("Got error retrieving playoff bracket information for team {} run {}", team.getTeamNumber(),
+                    nextRunNumber, e);
+        return new PerformanceRunsEndpoint.SelectTeamData(team, nextTableAndSide, nextRunNumber,
+                                                          String.valueOf(nextRunNumber));
+      }
+    }
+  }
+
+  /**
+   * @return the next table and side or the empty string if this cannot be
+   *         determined.
+   */
+  private static String getNextTableAndSide(final Connection connection,
+                                            final Tournament tournament,
+                                            final int teamNumber,
+                                            final int nextRunNumber) {
+    try {
+      if (nextRunNumber <= TournamentParameters.getNumSeedingRounds(connection, tournament.getTournamentID())) {
+        // no schedule, not going to know the table
+        return "";
+      } else {
+        // get information from the playoffs
+        return Playoff.getTableForRun(connection, tournament, teamNumber, nextRunNumber);
+      }
+    } catch (final SQLException e) {
+      LOGGER.warn("Got error retrieving next table and side for team {} run {}", teamNumber, nextRunNumber, e);
+      return "";
+    }
+  }
+
+  private static @Nullable PerformanceTime getNextPerformance(final @Nullable TournamentSchedule schedule,
+                                                              final int teamNumber,
+                                                              final int nextRunNumber) {
+    if (null == schedule) {
+      return null;
+    } else {
+      final @Nullable TeamScheduleInfo sched = schedule.getSchedInfoForTeam(teamNumber);
+      if (null == sched) {
+        return null;
+      } else {
+        // subtract 1 from the run number to get to a zero based index
+        final @Nullable PerformanceTime time = sched.enumerateRegularMatchPlayPerformances()//
+                                                    .filter(p -> p.getRight().longValue() == (nextRunNumber
+                                                        - 1))//
+                                                    .findFirst().map(Pair::getLeft).orElse(null);
+
+        return time;
+      }
     }
   }
 
@@ -245,9 +404,12 @@ public class PerformanceRunsEndpoint {
   public static final class PerformanceRunData {
 
     /**
+     * @param teamSelectData {@Link #getTeamSelectData()}
      * @param unverified {@link #getUnverified()}
      */
-    public PerformanceRunData(final Collection<UnverifiedRunData> unverified) {
+    public PerformanceRunData(final List<SelectTeamData> teamSelectData,
+                              final Collection<UnverifiedRunData> unverified) {
+      this.teamSelectData = teamSelectData;
       this.unverified = unverified;
     }
 
@@ -258,6 +420,15 @@ public class PerformanceRunsEndpoint {
      */
     public Collection<UnverifiedRunData> getUnverified() {
       return unverified;
+    }
+
+    private final List<SelectTeamData> teamSelectData;
+
+    /**
+     * @return information about the next runs for each team
+     */
+    public List<SelectTeamData> getTeamSelectData() {
+      return teamSelectData;
     }
 
   }
@@ -300,6 +471,92 @@ public class PerformanceRunsEndpoint {
      */
     public String getDisplay() {
       return display;
+    }
+  }
+
+  /**
+   * Data class for displaying team information on the web.
+   */
+  public static final class SelectTeamData {
+
+    private final TournamentTeam team;
+
+    private final @Nullable PerformanceTime nextPerformance;
+
+    private final int nextRunNumber;
+
+    private final String nextTableAndSide;
+
+    private final String runNumberDisplay;
+
+    SelectTeamData(final TournamentTeam team,
+                   final PerformanceTime nextPerformance,
+                   final int nextRunNumber) {
+      this.team = team;
+      this.nextPerformance = nextPerformance;
+      this.nextRunNumber = nextRunNumber;
+      this.nextTableAndSide = nextPerformance.getTableAndSide();
+      this.runNumberDisplay = String.valueOf(nextRunNumber);
+    }
+
+    SelectTeamData(final TournamentTeam team,
+                   final String nextTableAndSide,
+                   final int nextRunNumber,
+                   final String runNumberDisplay) {
+      this.team = team;
+      this.nextPerformance = null;
+      this.nextRunNumber = nextRunNumber;
+      this.nextTableAndSide = nextTableAndSide;
+      this.runNumberDisplay = runNumberDisplay;
+    }
+
+    /**
+     * @return what to display in the team selection box
+     */
+    public String getDisplayString() {
+      final String scheduleInfo;
+      if (null != nextPerformance) {
+        scheduleInfo = String.format(" @ %s", TournamentSchedule.formatTime(nextPerformance.getTime()));
+      } else {
+        scheduleInfo = "";
+      }
+      final String tableInfo;
+      if (nextTableAndSide.isEmpty()) {
+        tableInfo = "";
+      } else {
+        tableInfo = String.format(" - %s", nextTableAndSide);
+      }
+      return String.format("%d [%s] - %s%s%s", team.getTeamNumber(), team.getTrimmedTeamName(), runNumberDisplay,
+                           scheduleInfo, tableInfo);
+    }
+
+    /**
+     * @return team
+     */
+    public TournamentTeam getTeam() {
+      return team;
+    }
+
+    /**
+     * @return the next performance
+     */
+    public @Nullable PerformanceTime getNextPerformance() {
+      return nextPerformance;
+    }
+
+    /**
+     * @return next run number
+     */
+    public int getNextRunNumber() {
+      return nextRunNumber;
+    }
+
+    /**
+     * @return the table and side the team is performing on next, the empty string
+     *         if the table isn't known
+     */
+    public String getNextTableAndSide() {
+      return nextTableAndSide;
     }
   }
 }